requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
faiss-cpu==1.8.0
google-generativeai==0.8.0
pydantic==2.9.2
